added sound effects

+++ main.cpp
@@ -172,7 +172,7 @@ int main(){
 		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
 		//Getting delta time...
-		float deltaTime = getDeltaTime();
+		float deltaTime = (float)getDeltaTime();
 
 		// Camera matrix
 		camera.viewMatrix = lookAt(
